Process mail without matching attachments
Previously, if a message did not contain attachments, the program would
crash due to set -e. The status of the command did not get evaluated.
This commit adds set -e compatibility of the `mu extract` command and
only deletes messages that were processed, leaving unprocessed messages
in the mailbox for later processing if other image formats or message
processing is supported.