Extract all attachments
Extract all attachments from the mail. Select jpg etc. afterwards.  A bug is
encountered for extracting inline mime attachments where the file pattern did
not get applied. Then the wildcard file pattern in the for loop was not
evaluated. Using the save-attachments option here and processing the files from
the directory seems like an easy fix. This will also keep unprocessed files in
the temp directory for later analysis.